+++ crates/azure_types/src/policy_set_definition_name.rs
@@ -4,14 +4,13 @@ use compact_str::CompactString;
 use compact_str::ToCompactString;
 use serde::de::Error;
 use std::ops::Deref;
-use std::ops::DerefMut;
 use std::str::FromStr;
 use validator::Validate;
 
 /// https://learn.microsoft.com/en-us/azure/azure-resource-manager/management/resource-name-rules#microsoftauthorization
 #[derive(Debug, Clone, PartialEq, Eq, Hash, Validate, PartialOrd, Ord, Arbitrary)]
 pub struct PolicySetDefinitionName {
-    pub inner: CompactString,
+    inner: CompactString,
 }
 impl PolicySetDefinitionName {
     pub fn new(inner: impl Into<CompactString>) -> Self {
@@ -71,11 +70,6 @@ impl Deref for PolicySetDefinitionName {
 
     fn deref(&self) -> &Self::Target {
         &self.inner
-    }
-}
-impl DerefMut for PolicySetDefinitionName {
-    fn deref_mut(&mut self) -> &mut Self::Target {
-        &mut self.inner
     }
 }
 impl TryFrom<CompactString> for PolicySetDefinitionName {